Fixup after Rug api changes
- InstructionResponse -> Response
- New signature on ProjectEditor
- AddresableRugs now required for load
